+++ call_test.go
@@ -64,8 +64,9 @@ var _ = Describe("func Call()", func() {
 		startTime = time.Now()
 		buf = &fact.Buffer{}
 
-		test = New(app).Begin(
+		test = Begin(
 			t,
+			app,
 			WithStartTime(startTime),
 			WithOperationOptions(
 				engine.WithObserver(buf),