+++ testoption.go
@@ -21,6 +21,7 @@ func WithStartTime(t time.Time) TestOption {
 
 // testOptions is a container for the options set via TestOption values.
 type testOptions struct {
+	engineOptions    []engine.Option
 	operationOptions []engine.OperationOption
 	time             time.Time
 }
@@ -42,10 +43,18 @@ func newTestOptions(options []TestOption) *testOptions {
 	return ro
 }
 
+// WithEngineOptions returns a TestOption that applies optional settings to
+// the engine used by the test-runner.
+func WithEngineOptions(options ...engine.Option) TestOption {
+	return func(to *testOptions) {
+		to.engineOptions = append(to.engineOptions, options...)
+	}
+}
+
 // WithOperationOptions returns a TestOption that applies optional per-operation
 // settings when performing assertions.
 func WithOperationOptions(options ...engine.OperationOption) TestOption {
-	return func(ro *testOptions) {
-		ro.operationOptions = append(ro.operationOptions, options...)
+	return func(to *testOptions) {
+		to.operationOptions = append(to.operationOptions, options...)
 	}
 }